# Backend/dist/swagger.yaml
openapi: 3.0.3
info:
  title: Agencia Viagens API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /auth/login:
    post:
      summary: Login e obtenção de token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Token e role
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  role:
                    type: string
        '401':
          description: Credenciais inválidas
  /packages:
    get:
      summary: Listar pacotes
      parameters:
        - in: query
          name: tipo
          schema:
            type: string
            enum: [nacional, internacional]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pacotes
    post:
      summary: Criar pacote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TravelPackageInput'
      responses:
        '201':
          description: Pacote criado
  /packages/{id}:
    put:
      summary: Atualizar pacote
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TravelPackageInput'
      responses:
        '200':
          description: Pacote atualizado
    delete:
      summary: Remover pacote
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
      responses:
        '204':
          description: Removido
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TravelPackageInput:
      type: object
      required: [nome, preco, descricao, tipo]
      properties:
        nome:
          type: string
        preco:
          type: string
        descricao:
          type: string
        imagem:
          type: string
        tipo:
          type: string
          enum: [nacional, internacional]